Unittests will create a device if one is not present
